chore(money-input): move to packages
