+++ packages/components/inputs/money-input/src/currencies.json
