+++ packages/components/inputs/money-input/src/index.js
@@ -1,0 +1,3 @@
+export { default } from './money-input';
+
+export { default as version } from './version';
