+++ packages/components/inputs/money-input/src/money-input.js
@@ -19,7 +19,7 @@ import { DropdownIndicator } from '@commercetools-uikit/select-utils';
 import { FractionDigitsIcon } from '@commercetools-uikit/icons';
 import Constraints from '@commercetools-uikit/constraints';
 import { useToggleState } from '@commercetools-uikit/hooks';
-import createSelectStyles from '../../internals/create-select-styles';
+import createSelectStyles from '../../../../../src/components/internals/create-select-styles';
 import currencies from './currencies.json';
 import {
   getHighPrecisionWrapperStyles,
