+++ packages/components/inputs/money-input/src/money-input.visualroute.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { MoneyInput } from '@commercetools-frontend/ui-kit';
-import { Suite, Spec } from '../../../../test/percy';
+import { Suite, Spec } from '../../../../../test/percy';
 
 const value = {
   amount: '13.50',
