+++ packages/components/inputs/money-input/src/money-input.visualspec.js
