+++ packages/components/inputs/money-input/src/version.js
