--- src/components/inputs/money-input/index.js
@@ -1,1 +1,0 @@
-export { default } from './money-input';
